fix(checker): JSDoc import(...).A.B resolves a dotted @typedef qualified name (#17162)
A dotted JSDoc `@typedef {T} A.B` declares a *qualified* type name — the
module exports it under its full path `A.B`, not as a member literally named
`A`. tsz's string-based JSDoc import-type parser truncated the member to its
first identifier segment (`A`) and dropped `.B`. That truncation was latent
until #17139 added the value-only-export TS2694 check: with the tail gone, a
present qualified typedef looked like a missing member `A`, so a JSDoc
`import("./mod").A.B` reference mis-fired TS2694. Closes the two rows of
#17162 left after #17167 (jsDeclarationsImportNamespacedType, callbackCrossModule).

- parse_jsdoc_import_type now keeps every `.`-joined identifier segment of the
  member path (`A`, `A.B`, `A.B.C`), via a shared `leading_qualified_ident_len`
  scanner that also replaces the identical inline loop in
  jsdoc_nameless_typedef_host_name.
- resolve_jsdoc_import_type_reference resolves a qualified path against the
  module's typedef surface first (typedefs are stored under their full
  normalized name). Purely additive: single-segment members skip the branch, so
  the existing symbol -> typedef -> commonjs-expando -> TS2694 order is
  unchanged; a qualified non-typedef falls through to head-segment resolution,
  and TS2694 still names the first unresolved segment, matching tsc.
- Moved the two JSDoc import-reference resolvers into a new sibling
  resolution/import_reference.rs so name_resolution.rs stays under the 2000-line
  per-file budget (arch-size gate).

Adjacent tests: exact repro, three-level A.B.C, renamed binder, and two
negative controls (qualified missing head, single-segment missing) still TS2694.

Goal: hold
