+++ crates/tsz-checker/src/jsdoc/resolution/mod.rs
@@ -15,5 +15,6 @@
 //! - Import type resolution (`resolve_jsdoc_import_type_reference`)
 
 mod generic_typedef;
+mod import_reference;
 mod name_resolution;
 mod type_construction;
